feat: restructure config file layout
