+++ core/src/main/java/cc/azuramc/bedwars/listener/player/PlayerAFKListener.java
@@ -1,7 +1,7 @@
 package cc.azuramc.bedwars.listener.player;
 
 import cc.azuramc.bedwars.AzuraBedWars;
-import cc.azuramc.bedwars.config.object.PlayerConfig;
+import cc.azuramc.bedwars.config.object.SettingsConfig;
 import cc.azuramc.bedwars.game.GamePlayer;
 import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
@@ -19,9 +19,7 @@ import java.util.UUID;
  */
 public class PlayerAFKListener implements Listener {
 
-    private static final PlayerConfig.AFKCheck CONFIG = AzuraBedWars.getInstance().getPlayerConfig().getAfkCheck();
-
-    private static final int MAX_NO_MOVEMENT_TIME = CONFIG.getMaxNoMovementTime();
+    private static final SettingsConfig settingsConfig = AzuraBedWars.getInstance().getSettingsConfig();
 
     private static final Map<UUID, Long> AFK_LAST_MOVEMENT = new HashMap<>();
     private static int checkAFKTask;
@@ -53,7 +51,7 @@ public class PlayerAFKListener implements Listener {
                 long lastMove = AFK_LAST_MOVEMENT.getOrDefault(uuid, currentTime);
 
                 // 大于时间未移动
-                gamePlayer.setAfk(currentTime - lastMove >= MAX_NO_MOVEMENT_TIME * 1000L);
+                gamePlayer.setAfk(currentTime - lastMove >= settingsConfig.getMaxNoMovementTime() * 1000L);
             }
         }, 0L, 20L);
     }
